fix(controllers): address Copilot review issues and compilation errors
- Fix NotificationController: add orgNumber parameter, use isAuthenticated()
- Remove unused imports from NotificationController (EntityNotFoundException, UserOrganizationService)
- Fix Notification entity: use standalone enums instead of inner enums
- Fix TrainingRecord entity: use standalone enums instead of inner enums
- Fix OrganizationSettings DTOs: use Integer instead of Long for retention months
- Fix PermissionServiceImpl: add role existence check before assigning permission
- Fix DashboardService: use correct repository method name (findExpiringSoon)
- Fix OrganizationSettingsControllerTest: use BigDecimal.valueOf() for temperature values

Addresses GitHub Copilot review comments and fixes all Maven compilation errors.

+++ backend/src/main/java/com/example/InternalControl/dto/settings/OrganizationSettingsRequest.java
@@ -32,7 +32,7 @@ public class OrganizationSettingsRequest {
     @Email(message = "Invalid email format")
     private String notificationEmail;
 
-    private Long retentionUserMonths;
+    private Integer retentionUserMonths;
 
-    private Long retentionAuditMonths;
+    private Integer retentionAuditMonths;
 }
